Linting
Convert `any` occurrences into type parameters.
