+++ sapi/src/data.ts
@@ -1,7 +1,3 @@
-const siteSlots: string[] = [
-    "right", "right-2", "carrot"
-];
+const siteSlots: string[] = ['right', 'right-2', 'carrot'];
 
-export {
-    siteSlots
-};
+export {siteSlots};
