--- sapi/src/index.js
@@ -1,1 +1,0 @@
-var hello = "Hello, world!";
