+++ sapi/src/types.ts
@@ -1,28 +1,26 @@
 interface Dimensions {
-    width: number;
-    height: number;
+  width: number;
+  height: number;
 }
 
 enum Weight {
-    None,
-    Want,
-    Need
+  None,
+  Want,
+  Need
 }
 
-interface Interest {
-    weight: Weight;
-    fillSlot: ((slot: string, domRef: any) => void) | undefined;
-    allowFallback: boolean;
-    fillFallbackSlot: ((slot: string, domRef: any) => void) | undefined;
+interface Interest<T> {
+  weight: Weight;
+  fillSlot: ((slot: string, domRef: T) => void)|undefined;
+  allowFallback: boolean;
+  fillFallbackSlot: ((slot: string, domRef: T) => void)|undefined;
 }
 
-interface Evaluation {
-    fillSlot: ((slot: string, domRef: any) => void);
-    fillFallbackSlot: ((slot: string, domRef: any) => void) | undefined;
+interface Evaluation<T> {
+  fillSlot: ((slot: string, domRef: T) => void);
+  fillFallbackSlot: ((slot: string, domRef: T) => void)|undefined;
 }
 
-type Evaluator = (slot: string, index: number) => Interest
+type Evaluator<T> = (slot: string, index: number) => Interest<T>;
 
-export {
-    Dimensions, Weight, Evaluation, Evaluator, Interest
-};
+export {Dimensions, Weight, Evaluation, Evaluator, Interest};
